pyproject.toml
[tool.poetry]
name = "astral-arcade"
version = "0.1.0"
description = ""
authors = ["DDeenis <ddeenis46@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
flask-sqlalchemy = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
